+++ Runtime/Pooling/PoolConfig.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace BIMOS
+namespace KadenZombie8.Pooling
 {
     [CreateAssetMenu(fileName = "Pool Config", menuName = "BIMOS/Object Pool Config")]
     public class PoolConfig : ScriptableObject