+++ Runtime/Rig/BIMOSRig.cs
@@ -1,0 +1,23 @@
+using UnityEngine;
+
+namespace KadenZombie8.BIMOS.Rig
+{
+    public class BIMOSRig : MonoBehaviour
+    {
+        public static BIMOSRig Instance { get; private set; }
+
+        public ControllerRig ControllerRig;
+        public PhysicsRig PhysicsRig;
+        public AnimationRig AnimationRig;
+
+        private void Awake()
+        {
+            if (Instance != null && Instance != this)
+            {
+                Destroy(gameObject);
+                return;
+            }
+            Instance = this;
+        }
+    }
+}